Refactored a couple files, renamed a few files, and made sure the program and test-suite still work.
